Games dataset
This data would be used to create an app for users to find board and card games based on their genre, age range, length of time to play, or challenge level. The UI would be similar to the Netflix, and once the user chooses a game, they would see a description of that game and have access to a YouTube video on how to play. This dataset features the top 25 board games and top 25 card games from https://boardgames.com.